address code comments
