+++ internal/cache/archive/archive_test.go
@@ -10,7 +10,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/buildkite/agent/v3/internal/cache/internal/trace"
+	"github.com/buildkite/agent/v4/internal/cache/internal/trace"
 	"github.com/klauspost/compress/zip"
 )
